Add wrapTexts to d3-utils
